Refactor selectors.py by splitting it into smaller modules
The selectors module was large (and was only getting larger), so to
make the code easier to manage, methods module is created containing
clustering, dissimilarity, and partition submodules each containing
the corresponding methods. Test file for each module is created as well.
This also makes it easier when several people are working on the selection
methods.